+++ objects/stock.go
@@ -98,8 +98,8 @@ type StockQuote struct {
 	MarketCap            *float64 `json:"marketCap"`
 	PriceAvg50           float64  `json:"priceAvg50"`
 	PriceAvg200          float64  `json:"priceAvg200"`
-	Volume               int64    `json:"volume"`
-	AvgVolume            int64    `json:"avgVolume"`
+	Volume               float64  `json:"volume"`
+	AvgVolume            float64  `json:"avgVolume"`
 	Exchange             string   `json:"exchange"`
 	Open                 float64  `json:"open"`
 	PreviousClose        float64  `json:"previousClose"`
@@ -114,7 +114,7 @@ type StockQuote struct {
 type StockQuoteShot struct {
 	Symbol string  `json:"symbol"`
 	Price  float64 `json:"price"`
-	Volume int64   `json:"volume"`
+	Volume float64 `json:"volume"`
 }
 
 // RequestStockSearch ...
@@ -276,7 +276,7 @@ type StockCandle struct {
 	Low    float64 `json:"low"`
 	High   float64 `json:"high"`
 	Close  float64 `json:"close"`
-	Volume int64   `json:"volume"`
+	Volume float64 `json:"volume"`
 }
 
 // StockSymbolList ...
@@ -383,14 +383,14 @@ type SurvivorshipBiasFree struct {
 	High   float64 `json:"high"`
 	Low    float64 `json:"low"`
 	Close  float64 `json:"close"`
-	Volume int64   `json:"volume"`
+	Volume float64 `json:"volume"`
 	From   string  `json:"from"`
 }
 
 // OTCRealTimePrice ...
 type OTCRealTimePrice struct {
 	Symbol        string    `json:"symbol"`
-	Volume        int64     `json:"volume"`
+	Volume        float64   `json:"volume"`
 	PrevClose     float64   `json:"prevClose"`
 	High          float64   `json:"high"`
 	Low           float64   `json:"low"`
